chore: move feature flags to a separate file
